Added running
Теперь игрок может бегать.

+++ Source/After/Gameplay/Entity/Entity.h
@@ -56,10 +56,10 @@ protected:
 	bool bIsRunning;
 
 	UPROPERTY(BlueprintReadOnly, Category = "Moving")
-	int MovementX;
+	float MovementX;
 
 	UPROPERTY(BlueprintReadOnly, Category = "Moving")
-	int MovementY;
+	float MovementY;
 
 private:
